requirements.txt
pyserial
boto3
awscli
numpy